Pull out common code
